[fix] expiration of revoked tokens should not revive any unexpired revoked tokens
Previously, all revoked tokens for a user were stored under a single key
in the store, with automatic removal from the store of the entire key
based on the lifetime of the most recently revoked token.

Unexpired revoked tokens could be removed from the blacklist if the most
recently revoked token had a shorter lifetime, either because the most
recently revoked token had an iat earlier than other revoked tokens for
that user, or because the most recently revoked token was initially
given a shorter lifetime.

This fix stores each revoked token under a separate key and adds tests
for the above scenario as well as the new implementation of the purge
function, which is implemented as a single-per-user additional key.

This fix introduces a breaking change for custom stores. Each blacklist
check now requires two keys from the store, a per user check for any
purge, and a single check for revocation of the individual token.
In-memory, Redis, and Memcached stores have been modified to batch these
operations, with a substitution of an mget function rather than get;
custom stores must now implement mget rather than get as well.